refactor(native): move generation entry and streaming into its own module
`run_generation`, `project_native_outcomes`, `GenerationOptions` and its
helpers, the summarization calls, `stream_completion`, the subagent child-turn
helpers, and memory extraction move to `api_process_adapter/generation.rs`.

`GenerationOptions`, `summarize_turns`, `child_reply_turns`, and `run_child_turn`
keep the `pub(crate)` they already had — `anthropic_provider`,
`openai_compatible_provider`, `memory_extraction_gateway`,
`memory_selection_gateway`, `onepiece_planning`, `utility_delegation`, and
`subagent` all reach them through `api_process_adapter::…`, and `mod.rs`
re-exports them so those paths stay byte-identical.
